backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Portifolio backend"
authors = [
    {name = "Pedro Santos",email = "pedrohenriqueph004@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
